4/n - DEVREL-85: Cover payment api (#20)
# Fixes

* Covered missing attributes and endpoints of Payment API
* Added auto pagination to `listAllPayment()` method

### Checklist

- [x] I acknowledge that all my contributions will be made under the
project's license
- [x] I have made a material change to the repo (functionality, testing,
spelling, grammar)
- [x] I have titled the PR appropriately
- [x] I have updated my branch with the main branch
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added the necessary documentation about the functionality
in the appropriate .md file
- [ ] I have added inline documentation to the code I modified

If you have questions, create a GitHub Issue in this repository.

+++ trolley/Client.cs
@@ -3,8 +3,9 @@ using System.Text;
 using System.Net.Http;
 using Trolley.Exceptions;
 using Trolley.Types;
+using Trolley.Types.Supporting;
 using System.Security.Cryptography;
-using Newtonsoft.Json;
+using System.Threading.Tasks;
 
 namespace Trolley
 {
@@ -27,20 +28,20 @@ namespace Trolley
         /// </summary>
         /// <param name="endPoint">The api endPoint</param>
         /// <returns>The response</returns>
-        public string get(string endPoint)
+        public string Get(string endPoint)
         {
             string result = "";
             try {
-                httpClient = createRequest(endPoint, "GET");
+                httpClient = CreateRequest(endPoint, "GET");
                 HttpResponseMessage response = httpClient.GetAsync(endPoint).Result;
 
                 result = response.Content.ReadAsStringAsync().Result;
                 if ((int)response.StatusCode != 200)
                 {
-                    throwStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
+                    ThrowStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
                 }
             }
-            catch (System.Net.Http.HttpRequestException)
+            catch (HttpRequestException)
             {
                 throw new InvalidStatusCodeException(result);
             }
@@ -52,25 +53,48 @@ namespace Trolley
         /// Makes a POST request to API
         /// </summary>
         /// <param name="endPoint">The api endPoint</param>
-        /// <param name="stringBody">The request payload</param>
+        /// <param name="body">The request payload</param>
         /// <returns>The Response</returns>
-        public string post(string endPoint, ITrolleyMappable body)
+        public string Post(string endPoint, ITrolleyMappable body)
         {
-            body.IsMappable();
-            HttpContent jsonBody = convertBody(body.ToJson());
-            string result= "";
+            if (body.IsMappable())
+            {
+                return Post(endPoint, body.ToString());
+            }
+            else
+            {
+                throw new InvalidFieldException("Provided ITrolleyMappable object is not Mappable.");
+            }
+        }
+
+        /// <summary>
+        /// Post method with string body
+        /// </summary>
+        /// <param name="endPoint"></param>
+        /// <param name="body"></param>
+        /// <returns></returns>
+        /// <exception cref="InvalidStatusCodeException"></exception>
+        public string Post(string endPoint, string body)
+        {
+            if (body == null)
+            {
+                body = "";
+            }
+
+            HttpContent jsonBody = ConvertBody(body);
+            string result = "";
             try
             {
-                httpClient = createRequest(endPoint, "POST", body);
+                httpClient = CreateRequest(endPoint, "POST", null, body);
                 HttpResponseMessage response = httpClient.PostAsync(endPoint, jsonBody).Result;
                 result = response.Content.ReadAsStringAsync().Result;
                 if ((int)response.StatusCode != 200)
                 {
-                    throwStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
+                    ThrowStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
                 }
 
             }
-            catch (System.Net.Http.HttpRequestException)
+            catch (HttpRequestException)
             {
                 throw new InvalidStatusCodeException(result);
 
@@ -85,26 +109,26 @@ namespace Trolley
         /// <param name="endPoint">The api endPoint</param>
         /// <param name="stringBody">The request payload</param>
         /// <returns>The response</returns>
-        public string patch(string endPoint, ITrolleyMappable body)
+        public string Patch(string endPoint, ITrolleyMappable body)
         {
             body.IsMappable();
-            HttpContent jsonBody = convertBody(body.ToJson());
+            HttpContent jsonBody = ConvertBody(body.ToJson());
             string result = "";
             try
             {
-                httpClient = createRequest(endPoint, "PATCH", body);
+                httpClient = CreateRequest(endPoint, "PATCH", body);
 
                 var request = new HttpRequestMessage(new HttpMethod("PATCH"), endPoint) { Content = jsonBody };
-                System.Threading.Tasks.Task<HttpResponseMessage> responseTask = httpClient.SendAsync(request);
+                Task<HttpResponseMessage> responseTask = httpClient.SendAsync(request);
 
                 HttpResponseMessage response = responseTask.Result;
                 result = response.Content.ReadAsStringAsync().Result;
                 if ((int)response.StatusCode != 200)
                 {
-                    throwStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
+                    ThrowStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
                 }
             }
-            catch (System.Net.Http.HttpRequestException)
+            catch (HttpRequestException)
             {
                 throw new InvalidStatusCodeException(result);
             }
@@ -113,24 +137,41 @@ namespace Trolley
         }
 
         /// <summary>
-        /// Makes A DELETE request to API
+        /// Makes A DELETE request with body to API
         /// </summary>
         /// <param name="endPoint">The api endPoint</param>
+        /// <param name="body">The request body</param>
         /// <returns>The response</returns>
-        public string delete(string endPoint)
+        public string Delete(string endPoint, string body = null)
         {
             string result = "";
+            HttpResponseMessage response = null;
             try
             {
-                httpClient = createRequest(endPoint, "DELETE");
-                HttpResponseMessage response = httpClient.DeleteAsync(endPoint).Result;
+                if(body == null)
+                {
+                    httpClient = CreateRequest(endPoint, "DELETE");
+                    response = httpClient.DeleteAsync(endPoint).Result;
+                }
+                else
+                {
+                    httpClient = CreateRequest(endPoint, "DELETE", null, body);
+                    HttpContent jsonBody = ConvertBody(body);
+
+                    var request = new HttpRequestMessage(new HttpMethod("DELETE"), endPoint) { Content = jsonBody };
+                    Task<HttpResponseMessage> responseTask = httpClient.SendAsync(request);
+
+                    response = responseTask.Result;
+                }
+                
+                
                 result = response.Content.ReadAsStringAsync().Result;
                 if ((int)response.StatusCode != 200)
                 {
-                    throwStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
+                    ThrowStatusCodeException((int)response.StatusCode, response.Content.ReadAsStringAsync().Result);
                 }
             }
-            catch (System.Net.Http.HttpRequestException)
+            catch (HttpRequestException)
             {
                 throw new InvalidStatusCodeException(result);
             }
@@ -139,13 +180,13 @@ namespace Trolley
         }
 
 
-        private HttpContent convertBody(string body)
+        private HttpContent ConvertBody(string body)
         {
             HttpContent content = new StringContent(body, UTF8Encoding.UTF8, "application/json");
             return content;
         }
 
-        private HttpClient createRequest(string endPoint, string method, ITrolleyMappable body = null)
+        private HttpClient CreateRequest(string endPoint, string method, ITrolleyMappable body = null, string bodyStr = null)
         {
             try
             {
@@ -156,30 +197,30 @@ namespace Trolley
                 TimeSpan unixTicks = new TimeSpan(DateTime.UtcNow.Ticks) - epochTicks;
                 int unixTime = (int)unixTicks.TotalSeconds;
 
-                var authorization = generateAuthorization(unixTime + "", endPoint, method, body);
+                if (body != null)
+                {
+                    bodyStr = body.ToJson();
+                }
+
+                var authorization = GenerateAuthorization(unixTime + "", endPoint, method, bodyStr);
 
                 httpClient.DefaultRequestHeaders.Add("Authorization", authorization);
                 httpClient.DefaultRequestHeaders.Add("X-PR-Timestamp", unixTime + "");
-                httpClient.DefaultRequestHeaders.Add("Trolley-Source", "dotnet-sdk_1.0.10");
+                httpClient.DefaultRequestHeaders.Add("Trolley-Source", $"dotnet-sdk_{SemVer.MAJOR}.{SemVer.PATCH}.{SemVer.MINOR}");
                 return httpClient;
 
             }
-            catch (System.Net.Http.HttpRequestException e)
+            catch (HttpRequestException e)
             {
                 throw new InvalidStatusCodeException(e.Message);
             }
 
         }
 
-        private string generateAuthorization(string timeStamp, string endPoint, string method, ITrolleyMappable body)
+        private string GenerateAuthorization(string timeStamp, string endPoint, string method, string body)
         {
-            string newBody = "";
-            if (body != null)
-            {
-                newBody = body.ToJson();
-            }
 
-            string message = timeStamp + "\n" + method + "\n" + endPoint + "\n" + newBody + "\n";
+            string message = timeStamp + "\n" + method + "\n" + endPoint + "\n" + body + "\n";
 
             if (this.config.ApiSecret == null || this.config.ApiSecret == "")
             {
@@ -206,7 +247,7 @@ namespace Trolley
             return BitConverter.ToString(hashBytes).Replace("-", "").ToLower();
         }
 
-        private void throwStatusCodeException(int statusCode, string message = "")
+        private void ThrowStatusCodeException(int statusCode, string message = "")
         {
             switch (statusCode)
             {
